requirements.txt
opencv-python==4.10.0.84
tensorflow==2.12.0
numpy==1.23.5
pyserial==3.5